# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Possible Website API
  version: 1.0.0
  description: |
    REST API surface for the Possible Website marketplace. Endpoints follow JSON conventions and use bearer tokens for authentication.
servers:
  - url: https://api.possiblewebsite.test
    description: Production
  - url: http://localhost:4000
    description: Local development
paths:
...
  /api/orders:
    post:
      summary: Create an order
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/DirectOrderRequest'
                - $ref: '#/components/schemas/OfferOrderRequest'
                - $ref: '#/components/schemas/CartOrderRequest'
      responses:
        '201':
          description: Order created
...
components:
  schemas:
...
    DirectOrderRequest:
      type: object
      required: [type, buyerId, sellerId, listingId, amount]
      properties:
        type:
          const: direct
        buyerId:
          type: string
        sellerId:
          type: string
        listingId:
          type: string
        amount:
          type: number
        quantity:
          type: integer
        shippingAddressId:
          type: string
        billingAddressId:
          type: string
        offerId:
          type: string
    OfferOrderRequest:
      type: object
      required: [type, offerId]
      properties:
        type:
          const: offer
        offerId:
          type: string
        shippingAddressId:
          type: string
        billingAddressId:
          type: string
    CartOrderRequest:
      type: object
      required: [type, buyerId]
      properties:
        type:
          const: cart
        buyerId:
          type: string
...